Use the composer.json to register the test autoloading
Altering of the configuration of the autoloader generated by Composer is a
bad practice due to it disabling the OPCache-based performance optimizations
(configuring a separate instance of the class is fine).
Even though our testsuite is not concerned about this performance issue (the
OPCache optimization applies to PHP-FPM, not CLI), it is better to avoid
showing the bad practice.