Use PSR-18 and PSR-17 directly instead of HTTPlug
Projects using HTTPlug 2 will have an easy update, as HTTPlug 2
implements PSR-18 directly. For the factories, the mailer keeps a BC
layer supporting the HTTPlug factories rather than the PSR-17 one.
Projects using HTTPlug 1 won't be able to update Stampie until they
migrate (which is generally easy as the BC breaks between 1 and 2
are only for implementors of the interface, not for consumers).